requirements.txt
flet==0.22.*
google-generativeai